gateway-svc/requirements.txt
Flask==2.1.2
kafka-python
#python-schema-registry-client==2.4.0
requests==2.27.1